requirements.txt
torch>=2.1
torchvision>=0.16
pyyaml>=6.0
tqdm>=4.66
matplotlib>=3.8
scikit-learn>=1.3
# Optional (for DeiT/Swin via timm):
# timm>=0.9.12
